requirements.txt
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
ccxt>=4.3.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
pandas-ta>=0.3.14b
requests>=2.31.0
python-dotenv>=1.0.0
fredapi>=0.5.1
scipy>=1.11.0
pytz>=2024.1
